requirements.txt
email-validator==2.2.0
fastapi==0.114.1
pydantic==2.9.1
pytest==8.3.3
python-dotenv==1.0.1
uvicorn==0.30.6
